spec: move test model classes to spec_helper
