+++ spec/model_spec.rb
@@ -1,57 +1,7 @@
 require 'spec_helper'
 
 module RASN1
-
-  class ModelTest < Model
-    sequence :record,
-             content: [integer(:id),
-                       integer(:room, implicit: 0, optional: true),
-                       integer(:house, explicit: 1, default: 0)]
-  end
-
-  class ModelTest2 < Model
-    sequence :record2,
-             content: [boolean(:rented),
-                       model(:a_record, ModelTest)]
-  end
-
-  class ModelTest3 < ModelTest
-    root_options implicit: 4
-  end
-
-  class OfModel < Model
-    sequence_of :seqof, ModelTest
-  end
-
-  class SuperOfModel < Model
-    sequence :super,
-             content: [model(:of, OfModel)]
-  end
-
-  class VoidSeq < Model
-    sequence :voidseq
-  end
-
-  class VoidSeq2 < Model
-    sequence :voidseq2,
-             content: [boolean(:bool), sequence(:seq)]
-  end
-
-  class ModelChoice < Model
-    choice :choice,
-           content: [integer(:id),
-                     model(:a_record, ModelTest)]
-  end
-
-  class ExplicitTaggedSeq < Model
-    sequence :seq, explicit: 0, class: :application,
-             content: [integer(:id), integer(:extern_id)]
-  end
-
-  class ModelExplicitBitString < Model
-    sequence :bit_string,
-             content: [bit_string(:flags, explicit: 0, constructed: true, bit_length: 32)]
-  end
+  include TestModel
 
   SIMPLE_VALUE = "\x30\x0e\x02\x03\x01\x00\x01\x80\x01\x2b\x81\x04\x02\x02\x12\x34".force_encoding('BINARY')
   OPTIONAL_VALUE = "\x30\x0b\x02\x03\x01\x00\x01\x81\x04\x02\x02\x12\x34".force_encoding('BINARY')
