+++ spec/types_spec.rb
@@ -23,9 +23,9 @@ module RASN1
     describe '.constructed' do
       it 'gives all subclasses of Constructed' do
         expect(Types.constructed).to include(Types::Sequence,
-                                              Types::SequenceOf,
-                                              Types::Set,
-                                              Types::SetOf)
+                                             Types::SequenceOf,
+                                             Types::Set,
+                                             Types::SetOf)
       end
     end
   end
